refactor(routing): split the event router into pipeline, settlement, gate, and dead-letter units
EventRouter shrinks to stream orchestration. The per-subscription closure
style stays intact (context built once, hot path never touches this), the
concurrency limiter moves into ConcurrencyGate, settlement into its own
factory, and the whole DLQ flow into DeadLetterCapture. Route-pipeline
benchmark holds at baseline (sync ~3.16M ops/s, async ~0.8M ops/s).
